Multiple Lights
# Fixed the bug that was haunting this repository

Importer.cpp was (shamefully) *heavily inspired* by a blogpost, it was, however, completely incorrect, I guess that this is a lesson to always double-check code from the internet.

Importer.cpp now correctly translates tiny_obj_loader format into VBOs and EBOs ready to be used in a model class, this results in correct lightning rendering in our application!

# Multiple Lights

Now that the object importing is fixed, the application finally supports multiple movable point lights in the scene.