Normal Matrix Bugfix
A bug was occurring due to GLM's mat3 alignment which inserted a padding byte every three bytes, this caused some entries of the normal matrix to be 0.0f, leading to incorrect calculations.

This has been patched by passing a mat4 around and casting it in the shader, it is a preferrable alternative to eliminating the padding due to the fact that the padding helps the compiler with SIMD instruction s on the CPU side.